DOMCore/tests/submissions/Ms2ger/DOMImplementation-createDocumentType.html

<!doctype html>
<title>DOMImplementation.createDocumentType(qualifiedName, publicId, systemId)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var tests = [
        ["foo", "", "", null],
        ["1foo", "", "", "INVALID_CHARACTER_ERR"],
        ["foo1", "", "", null],
        ["f1oo", "", "", null],
        ["@foo", "", "", "INVALID_CHARACTER_ERR"],
        ["foo@", "", "", "INVALID_CHARACTER_ERR"],
        ["f@oo", "", "", "INVALID_CHARACTER_ERR"],
        ["f:oo", "", "", null],
        [":foo", "", "", "NAMESPACE_ERR"],
        ["foo:", "", "", "NAMESPACE_ERR"],
        ["foo", "foo", "", null],
        ["foo", "", "foo", null],
        ["foo", "f'oo", "", null],
        ["foo", "", "f'oo", null],
        ["foo", 'f"oo', "", null],
        ["foo", "", 'f"oo', null],
        ["foo", "f'o\"o", "", null],
        ["foo", "", "f'o\"o", null],
        ["foo", "foo>", "", null],
        ["foo", "", "foo>", null]
     ]

  tests.forEach(function(t) {
    var qualifiedName = t[0], publicId = t[1], systemId = t[2], expected = t[3]
    test(function() {
      if (expected) {
        assert_throws(expected, function() {
          document.implementation.createDocumentType(qualifiedName, publicId, systemId)
        })
      } else {
        var doctype = document.implementation.createDocumentType(qualifiedName, publicId, systemId);
        assert_equals(doctype.name, qualifiedName)
        assert_equals(doctype.publicId, publicId)
        assert_equals(doctype.systemId, systemId)
      }
    }, "createDocumentType(" + format_value(qualifiedName) + ", " + format_value(publicId) + ", " + format_value(systemId) + ")");
  });
})
</script>
